+++ vogue/server/endpoints/bioinfo/__init__.py
@@ -1,0 +1,4 @@
+from .covid import covid_blueprint
+from .mip import mip_blueprint
+from .micro import micro_blueprint
+from .cancer import cancer_blueprint
