+++ vogue/server/templates/genotype_plate.html
@@ -10,7 +10,7 @@
                 <div class="card-body">
                     <div class="row">
                         <div class="btn mb-2 mb-md-0">
-                            <form action="{{ url_for('server.genotype_plate') }}" method="POST">
+                            <form action="{{ url_for('genotype.genotype_plate') }}" method="POST">
                                 <select name="plate_id" class="form-control form-control-sm"
                                     placeholder=".form-control-sm" required onchange="this.form.submit()">
                                     {% for plate in plates %}
