+++ vogue/server/templates/header_section.html
@@ -2,7 +2,7 @@
 <div class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pt-3 pb-2 mb-3 border-bottom">
   <h1 class="h2">{{header}}</h1>
   <div class="btn-toolbar mb-2 mb-md-0">
-    <form action="{{ url_for('server.index')}}" method="POST">
+    <form action="{{ url_for('home.index')}}" method="POST">
       <input type="hidden" name="page" value="{{page_id}}">
       {% if years %}
       <select name="year" class="form-control" required onchange="this.form.submit()">
