+++ vogue/server/templates/microsalt_st_time.html
@@ -10,7 +10,7 @@
                 <div class="card-body">
                     <div class="row">
                         <div class="btn mb-2 mb-md-0">
-                            <form action="{{ url_for('server.microsalt_st_time' , year=year_of_interest) }}"
+                            <form action="{{ url_for('micro.st_time' , year=year_of_interest) }}"
                                 method="POST">
                                 <select id="strain" name="strain" class="form-control form-control-sm"
                                     placeholder=".form-control-sm" required onchange="this.form.submit()">
