+++ vogue/server/templates/mip_dna_picard.html
@@ -10,7 +10,7 @@
                 <div class="card-body">
                     <div class="row">
                         <div class="btn mb-3 mb-md-0">
-                            <form action="{{ url_for('server.mip_dna_picard' , year=year_of_interest) }}" method="POST">
+                            <form action="{{ url_for('mip.dna_picard' , year=year_of_interest) }}" method="POST">
                                 <label class="form-text text-muted" for="X_axis">Select X axis metric:</label>
                                 <select id="X_axis" name="X_axis" class="form-control form-control-sm"
                                     placeholder=".form-control-sm" required onchange="this.form.submit()">
