+++ vogue/server/templates/reagent_labels.html
@@ -10,7 +10,7 @@
                 <div class="card-body">
                     <div class="row">
                             <div class="btn mb-2 mb-md-0">
-                                <form action="{{ url_for('server.index')}}" method="POST">
+                                <form action="{{ url_for('home.index')}}" method="POST">
                                     <input type="hidden" name="page" value="{{page_id}}"> 
                                         <select id="index_category" name="index_category" class="form-control form-control-sm"
                                             placeholder=".form-control-sm" required onchange="this.form.submit()">
@@ -79,7 +79,7 @@
     tooltip: {
         style: {pointerEvents: 'all'},
         useHTML: true,
-        pointFormat: 'Index performance: <b>{point.y:.1f}</b><br>Average from <b>{point.nr_runs}</b> runs. <br><a href="{{url_for('server.reagent_label', index_category=index_category.replace(' ','_'), reagent_label='')}}{point.url}">{point.url}</a>'
+        pointFormat: 'Index performance: <b>{point.y:.1f}</b><br>Average from <b>{point.nr_runs}</b> runs. <br><a href="{{url_for('index.reagent_label', index_category=index_category.replace(' ','_'), reagent_label='')}}{point.url}">{point.url}</a>'
 
     },
     series: [{
