+++ vogue/server/templates/runs.html
@@ -10,7 +10,7 @@
                 <div class="card-body">
                     <div class="row">
                         <div class="btn mb-2 mb-md-0">
-                            <form action="{{ url_for('server.runs' , year=year_of_interest) }}" method="POST">
+                            <form action="{{ url_for('sequencing.runs' , year=year_of_interest) }}" method="POST">
                                 <select id="metric" name="metric" class="form-control form-control-sm"
                                     placeholder=".form-control-sm" required onchange="this.form.submit()">
                                     <option value="">Select metric...</option>
