+++ vogue/server/utils/bioinfo/__init__.py
@@ -1,0 +1,1 @@
+from .covid import get_qc
